Hovering over sessions now displays the bubble
Clicking on each session in the session pane also makes the bubbles
persistently displayed until you click a new session in the list, a
pin, or the map.

I also removed the drop down animation on the course picker so that it
could be statically placed to avoid bubbles being displayed off screen.