Fixed non-registered workflow
Sessions refactor also broke non-registered workflow in several ways.
This has been fixed now so that the same experience is given to both
registered and non-registered users.

One interesting thing to note is the fact that filtering for registered
users takes longer than for non-registered users due to the fact that
we currently wait for the filtering db call to be made before actually
filtering. Not sure if this is an issue or not, but it can be fixed if
so.