I believe we can now view hosting and attending sessions.
It took me a while to come up with the best way to do this, but using
filters when displaying only seemed to be the best way to do this.
StateService now keeps track of the filter that’s being employed (based
on which tab is selected) and the UI throughout the application is
updated accordingly.